Reimplement `colophone.js` to use webpack's 'val-loader'. Tests fail...
The 'val-loader' executes a function to generate the module exports.
Conceptually this is exactly what `colophone.js` needs.

However, to make TypeScript happy, the 'val-loader' result must be a
function. 'JsDoc' syntax annotates the return value so TypeScript
will accept it.

Tests still fail. Making this work with karma requires more study.
Karma does not use webpack and its loaders.
One possible approach is to implement 'colophone' as a service,
and then use a mock for testing.
Using a service may be cleaner overall, too.